backend/requirements.txt
alembic==1.14.1
aiohttp==3.11.11
aioboto3==13.4.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
passlib==1.7.4
click==8.1.8
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.7
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
pyasn1==0.6.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.2
typing_extensions==4.12.2
uvicorn==0.34.0
requests==2.32.3
python-magic==0.4.27
celery==5.3.4